# swagger.yaml
swagger: "2.0"
info:
  description: "Travel Booking Portal API (Microservices)"
  version: "1.0.0"
  title: "Travel Booking Portal"
host: "localhost:8080"
basePath: "/"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /booking/{bookingID}:
    put:
      operationId: "UpdateBooking"
      parameters:
        - in: "path"
          name: "bookingID"
          description: "Booking object to update"
          required: true
          type: "integer"
        - in: "body"
          name: "booking"
          description: "Booking object to update"
          required: true
          schema:
            $ref: "#/definitions/Booking"
      responses:
        200:
          description: " updated successfully"
          schema:
            $ref: "#/definitions/Booking"
        201:
          description: "Booking updated successfully"
          schema:
            $ref: "#/definitions/Booking"
        400:
          description: "Bad Request - The request is invalid or contains errors"
        409:
          description: "Booking already exists"
        404:
          description: "Booking not found"
        500:
          description: "Internal Server Error"
    get:
      operationId: "GetBooking"
      parameters:
        - name: "bookingID"
          in: "path"
          description: "ID of the booking to retrieve"
          required: true
          type: "integer"
      responses:
        200:
          description: "Booking retrieved successfully"
          schema:
            $ref: "#/definitions/Booking"
        404:
          description: "Booking not found"
        500:
          description: "Internal Server Error"
    delete:
      operationId: "DeleteBooking"
      parameters:
        - name: "bookingID"
          in: "path"
          description: "ID of the booking to delete"
          required: true
          type: "integer"
      responses:
        400:
          description: "Bad Request - The request is invalid or contains errors"
        204:
          description: "Booking deleted successfully"
        404:
          description: "Booking not found"
        500:
          description: "Internal Server Error"
...
definitions:
...
  Booking:
    type: "object"
    properties:
      id:
        type: "integer"
      user_id:
        type: "integer"
      service_type:
        type: "string"
      service_id:
        type: "integer"
      from_date:
        type: "string"
        format: "date-time"
      to_date:
        type: "string"
        format: "date-time"
      confirmed:
        type: "boolean"
      penalty:
        type: "boolean"
      paid:
        type: "boolean"